Break out VisualColorPickersInput.svelte
