+++ frontend/src/components/floating-menus/ColorPicker.svelte
@@ -1,15 +1,15 @@
 <script lang="ts">
-	import { getContext, onDestroy, createEventDispatcher, tick } from "svelte";
-	import FloatingMenu, { preventEscapeClosingParentFloatingMenu } from "/src/components/layout/FloatingMenu.svelte";
+	import { createEventDispatcher, tick } from "svelte";
+	import FloatingMenu from "/src/components/layout/FloatingMenu.svelte";
 	import LayoutCol from "/src/components/layout/LayoutCol.svelte";
 	import LayoutRow from "/src/components/layout/LayoutRow.svelte";
 	import IconButton from "/src/components/widgets/buttons/IconButton.svelte";
 	import NumberInput from "/src/components/widgets/inputs/NumberInput.svelte";
 	import SpectrumInput, { MAX_MIDPOINT, MIN_MIDPOINT } from "/src/components/widgets/inputs/SpectrumInput.svelte";
 	import TextInput from "/src/components/widgets/inputs/TextInput.svelte";
+	import VisualColorPickersInput from "/src/components/widgets/inputs/VisualColorPickersInput.svelte";
 	import Separator from "/src/components/widgets/labels/Separator.svelte";
 	import TextLabel from "/src/components/widgets/labels/TextLabel.svelte";
-	import type { TooltipStore } from "/src/stores/tooltip";
 	import type { HSV, RGB } from "/src/utility-functions/colors";
 	import {
 		contrastingOutlineFactor,
@@ -53,7 +53,6 @@
 	];
 
 	const dispatch = createEventDispatcher<{ colorOrGradient: FillChoice; startHistoryTransaction: undefined; commitHistoryTransaction: undefined }>();
-	const tooltip = getContext<TooltipStore>("tooltip");
 
 	export let colorOrGradient: FillChoice;
 	export let allowNone = false;
@@ -89,19 +88,8 @@
 	let oldAlpha = colorForHSVA ? colorForHSVA.alpha : 1;
 	let oldIsNone = hsvOrNone === undefined;
 	// Transient state
-	let draggingPickerTrack: HTMLDivElement | undefined = undefined;
 	let strayCloses = true;
 	let gradientSpectrumDragging = false;
-	let shiftPressed = false;
-	let alignedAxis: "saturation" | "value" | undefined = undefined;
-	let hueBeforeDrag = 0;
-	let saturationBeforeDrag = 0;
-	let valueBeforeDrag = 0;
-	let alphaBeforeDrag = 0;
-	let saturationStartOfAxisAlign: number | undefined = undefined;
-	let valueStartOfAxisAlign: number | undefined = undefined;
-	let saturationRestoreWhenShiftReleased: number | undefined = undefined;
-	let valueRestoreWhenShiftReleased: number | undefined = undefined;
 
 	let self: FloatingMenu | undefined;
 	let hexCodeInputWidget: TextInput | undefined;
@@ -163,145 +151,11 @@
 		isNone = false;
 	}
 
-	function onPointerDown(e: PointerEvent) {
-		if (disabled) return;
-
-		const target = e.target instanceof HTMLElement ? e.target : undefined;
-		draggingPickerTrack = target?.closest("[data-saturation-value-picker], [data-hue-picker], [data-alpha-picker]") || undefined;
-
-		hueBeforeDrag = hue;
-		saturationBeforeDrag = saturation;
-		valueBeforeDrag = value;
-		alphaBeforeDrag = alpha;
-
-		saturationStartOfAxisAlign = undefined;
-		valueStartOfAxisAlign = undefined;
-
-		addEvents();
-
-		onPointerMove(e);
-	}
-
-	function onPointerMove(e: PointerEvent) {
-		// Just in case the mouseup event is lost
-		if (e.buttons === 0) removeEvents();
-
-		if (draggingPickerTrack?.hasAttribute("data-saturation-value-picker")) {
-			const rectangle = draggingPickerTrack.getBoundingClientRect();
-
-			saturation = clamp((e.clientX - rectangle.left) / rectangle.width, 0, 1);
-			value = clamp(1 - (e.clientY - rectangle.top) / rectangle.height, 0, 1);
-			strayCloses = false;
-
-			if (shiftPressed) updateAxisLock();
-		} else if (draggingPickerTrack?.hasAttribute("data-hue-picker")) {
-			const rectangle = draggingPickerTrack.getBoundingClientRect();
-
-			hue = clamp(1 - (e.clientY - rectangle.top) / rectangle.height, 0, 1);
-			strayCloses = false;
-		} else if (draggingPickerTrack?.hasAttribute("data-alpha-picker")) {
-			const rectangle = draggingPickerTrack.getBoundingClientRect();
-
-			alpha = clamp(1 - (e.clientY - rectangle.top) / rectangle.height, 0, 1);
-			strayCloses = false;
-		}
-
-		const color = createColorFromHSVA(hue, saturation, value, alpha);
-		setColor(color);
-
-		if (!e.shiftKey) {
-			shiftPressed = false;
-			alignedAxis = undefined;
-		} else if (!shiftPressed && draggingPickerTrack) {
-			shiftPressed = true;
-			saturationStartOfAxisAlign = saturationBeforeDrag;
-			valueStartOfAxisAlign = valueBeforeDrag;
-		}
-	}
-
-	function onPointerUp() {
-		if (draggingPickerTrack) dispatch("commitHistoryTransaction");
-		removeEvents();
-	}
-
-	function onMouseDown(e: MouseEvent) {
-		const BUTTONS_RIGHT = 0b0000_0010;
-		if (e.buttons & BUTTONS_RIGHT) abortDrag();
-	}
-
-	function onKeyDown(e: KeyboardEvent) {
-		if (e.key === "Escape") {
-			const element = self?.div();
-			if (element) preventEscapeClosingParentFloatingMenu(element);
-
-			abortDrag();
-		}
-	}
-
-	function onKeyUp(e: KeyboardEvent) {
-		if (e.key === "Shift") {
-			shiftPressed = false;
-			alignedAxis = undefined;
-
-			if (saturationRestoreWhenShiftReleased !== undefined && valueRestoreWhenShiftReleased !== undefined) {
-				saturation = saturationRestoreWhenShiftReleased;
-				value = valueRestoreWhenShiftReleased;
-
-				const color = createColorFromHSVA(hue, saturation, value, alpha);
-				setColor(color);
-			}
-		}
-	}
-
-	function addEvents() {
-		document.addEventListener("pointermove", onPointerMove);
-		document.addEventListener("pointerup", onPointerUp);
-		document.addEventListener("mousedown", onMouseDown);
-		document.addEventListener("keydown", onKeyDown);
-		document.addEventListener("keyup", onKeyUp);
-
-		dispatch("startHistoryTransaction");
-	}
-
-	function removeEvents() {
-		draggingPickerTrack = undefined;
-		// The setTimeout is necessary to prevent the FloatingMenu's `escapeCloses` from becoming true immediately upon pressing the Escape key, and thus closing
-		setTimeout(() => (strayCloses = true), 0);
-		shiftPressed = false;
-		alignedAxis = undefined;
-
-		document.removeEventListener("pointermove", onPointerMove);
-		document.removeEventListener("pointerup", onPointerUp);
-		document.removeEventListener("mousedown", onMouseDown);
-		document.removeEventListener("keydown", onKeyDown);
-		document.removeEventListener("keyup", onKeyUp);
-	}
-
-	function updateAxisLock() {
-		if (!saturationStartOfAxisAlign || !valueStartOfAxisAlign) return;
-
-		const deltaSaturation = saturation - saturationStartOfAxisAlign;
-		const deltaValue = value - valueStartOfAxisAlign;
-
-		saturationRestoreWhenShiftReleased = saturation;
-		valueRestoreWhenShiftReleased = value;
-
-		if (Math.abs(deltaSaturation) < Math.abs(deltaValue)) {
-			alignedAxis = "saturation";
-			saturation = saturationStartOfAxisAlign;
-		} else {
-			alignedAxis = "value";
-			value = valueStartOfAxisAlign;
-		}
-	}
-
-	function abortDrag() {
-		removeEvents();
-
-		hue = hueBeforeDrag;
-		saturation = saturationBeforeDrag;
-		value = valueBeforeDrag;
-		alpha = alphaBeforeDrag;
+	function onVisualUpdate({ detail }: CustomEvent<{ hue: number; saturation: number; value: number; alpha: number }>) {
+		hue = detail.hue;
+		saturation = detail.saturation;
+		value = detail.value;
+		alpha = detail.alpha;
 
 		const color = createColorFromHSVA(hue, saturation, value, alpha);
 		setColor(color);
@@ -434,17 +288,9 @@
 		setOldHSVA(hsv.h, hsv.s, hsv.v, color.alpha, false);
 	}
 
-	function clamp(value: number, min = 0, max = 1): number {
-		return Math.max(min, Math.min(value, max));
-	}
-
 	export function div(): HTMLDivElement | undefined {
 		return self?.div();
 	}
-
-	onDestroy(() => {
-		removeEvents();
-	});
 </script>
 
 <FloatingMenu class="color-picker" classes={{ disabled }} {open} on:open {strayCloses} escapeCloses={strayCloses && !gradientSpectrumDragging} {direction} type="Popover" bind:this={self}>
@@ -464,58 +310,19 @@
 		{@const saturationDescription = "The vividness from grayscale to full color."}
 		{@const valueDescription = "The brightness from black to full color."}
 		<LayoutCol class="pickers-and-gradient">
-			<LayoutRow class="pickers">
-				<LayoutCol
-					class="saturation-value-picker"
-					data-tooltip-label="Saturation and Value"
-					data-tooltip-description={`To move only along the saturation (X) or value (Y) axis, perform the shortcut shown.${disabled ? "\n\nDisabled (read-only)." : ""}`}
-					data-tooltip-shortcut={$tooltip.shiftClickShortcut?.shortcut ? JSON.stringify($tooltip.shiftClickShortcut.shortcut) : undefined}
-					on:pointerdown={onPointerDown}
-					data-saturation-value-picker
-				>
-					{#if alignedAxis}
-						<div
-							class="selection-circle-axis-snap-line"
-							style:width={alignedAxis === "value" ? "100%" : undefined}
-							style:height={alignedAxis === "saturation" ? "100%" : undefined}
-							style:top={alignedAxis === "value" ? `${(1 - value) * 100}%` : undefined}
-							style:left={alignedAxis === "saturation" ? `${saturation * 100}%` : undefined}
-						></div>
-						<div
-							class="selection-circle-axis-snap-line"
-							style:width={alignedAxis === "saturation" ? "100%" : undefined}
-							style:height={alignedAxis === "value" ? "100%" : undefined}
-							style:top={alignedAxis === "saturation" ? `${(1 - valueBeforeDrag) * 100}%` : undefined}
-							style:left={alignedAxis === "value" ? `${saturationBeforeDrag * 100}%` : undefined}
-						></div>
-					{/if}
-					{#if !isNone}
-						<div class="selection-circle" style:top={`${(1 - value) * 100}%`} style:left={`${saturation * 100}%`}></div>
-					{/if}
-				</LayoutCol>
-				<LayoutCol
-					class="hue-picker"
-					data-tooltip-label="Hue"
-					data-tooltip-description={`${hueDescription}${disabled ? "\n\nDisabled (read-only)." : ""}`}
-					on:pointerdown={onPointerDown}
-					data-hue-picker
-				>
-					{#if !isNone}
-						<div class="selection-needle" style:top={`${(1 - hue) * 100}%`}></div>
-					{/if}
-				</LayoutCol>
-				<LayoutCol
-					class="alpha-picker"
-					data-tooltip-label="Alpha"
-					data-tooltip-description={`The level of translucency.${disabled ? "\n\nDisabled (read-only)." : ""}`}
-					on:pointerdown={onPointerDown}
-					data-alpha-picker
-				>
-					{#if !isNone}
-						<div class="selection-needle" style:top={`${(1 - alpha) * 100}%`}></div>
-					{/if}
-				</LayoutCol>
-			</LayoutRow>
+			<VisualColorPickersInput
+				{hue}
+				{saturation}
+				{value}
+				{alpha}
+				{isNone}
+				{disabled}
+				getFloatingMenuElement={() => self?.div()}
+				on:update={onVisualUpdate}
+				on:startHistoryTransaction={() => dispatch("startHistoryTransaction")}
+				on:commitHistoryTransaction={() => dispatch("commitHistoryTransaction")}
+				on:dragStateChange={({ detail }) => (strayCloses = !detail)}
+			/>
 			{#if gradient}
 				<LayoutRow class="gradient">
 					<SpectrumInput
@@ -735,124 +542,8 @@
 <style lang="scss">
 	.color-picker {
 		--widget-height: 24px;
-		--picker-size: 256px;
-		--picker-circle-radius: 6px;
 
 		.pickers-and-gradient {
-			.pickers {
-				.saturation-value-picker {
-					width: var(--picker-size);
-					background-blend-mode: multiply;
-					background: linear-gradient(to bottom, #ffffff, #000000), linear-gradient(to right, #ffffff, var(--hue-color));
-					position: relative;
-				}
-
-				.saturation-value-picker,
-				.hue-picker,
-				.alpha-picker {
-					height: var(--picker-size);
-					border-radius: 2px;
-					position: relative;
-					overflow: hidden;
-				}
-
-				.hue-picker,
-				.alpha-picker {
-					width: 24px;
-					margin-left: 8px;
-					position: relative;
-				}
-
-				.hue-picker {
-					--selection-needle-color: var(--hue-color-contrasting);
-					background-blend-mode: screen;
-					background:
-						// Reds
-						linear-gradient(to top, #ff0000ff calc(100% / 6), #ff000000 calc(200% / 6), #ff000000 calc(400% / 6), #ff0000ff calc(500% / 6)),
-						// Greens
-						linear-gradient(to top, #00ff0000 0%, #00ff00ff calc(100% / 6), #00ff00ff 50%, #00ff0000 calc(400% / 6)),
-						// Blues
-						linear-gradient(to top, #0000ff00 calc(200% / 6), #0000ffff 50%, #0000ffff calc(500% / 6), #0000ff00 100%);
-				}
-
-				.alpha-picker {
-					--selection-needle-color: var(--new-color-contrasting);
-					background-image: linear-gradient(to bottom, var(--opaque-color), transparent), var(--color-transparent-checkered-background);
-					background-size:
-						100% 100%,
-						var(--color-transparent-checkered-background-size);
-					background-position:
-						0 0,
-						var(--color-transparent-checkered-background-position);
-					background-repeat: no-repeat, var(--color-transparent-checkered-background-repeat);
-				}
-
-				.selection-circle {
-					pointer-events: none;
-					position: absolute;
-					left: 0;
-					top: 0;
-					width: 0;
-					height: 0;
-
-					&::after {
-						content: "";
-						display: block;
-						position: relative;
-						left: calc(-1 * var(--picker-circle-radius));
-						top: calc(-1 * var(--picker-circle-radius));
-						width: calc(var(--picker-circle-radius) * 2 + 1px);
-						height: calc(var(--picker-circle-radius) * 2 + 1px);
-						border-radius: 50%;
-						border: 2px solid var(--opaque-color-contrasting);
-						background: var(--opaque-color);
-						box-sizing: border-box;
-					}
-				}
-
-				.selection-circle-axis-snap-line {
-					pointer-events: none;
-					position: absolute;
-					width: 1px;
-					height: 1px;
-					top: 0;
-					left: 0;
-					background: var(--opaque-color-contrasting);
-
-					+ .selection-circle-axis-snap-line {
-						opacity: 0.25;
-					}
-				}
-
-				.selection-needle {
-					pointer-events: none;
-					position: absolute;
-					top: 0;
-					width: 100%;
-					height: 0;
-
-					&::before {
-						content: "";
-						position: absolute;
-						top: -4px;
-						left: 0;
-						border-style: solid;
-						border-width: 4px 0 4px 4px;
-						border-color: transparent transparent transparent var(--selection-needle-color);
-					}
-
-					&::after {
-						content: "";
-						position: absolute;
-						top: -4px;
-						right: 0;
-						border-style: solid;
-						border-width: 4px 4px 4px 0;
-						border-color: transparent var(--selection-needle-color) transparent transparent;
-					}
-				}
-			}
-
 			.gradient {
 				margin-top: 16px;
 
@@ -1057,7 +748,6 @@
 			}
 		}
 
-		&.disabled .pickers-and-gradient .pickers :is(.saturation-value-picker, .hue-picker, .alpha-picker),
 		&.disabled .details .preset-color,
 		&.disabled .details .choice-preview {
 			transition: opacity 0.1s;
@@ -1071,6 +761,4 @@
 			background: var(--pure-color-gray);
 		}
 	}
-
-	// paddingpaddingpaddingpaddingpaddingpaddingpaddingpaddingpaddingpaddingpaddingpaddingpaddingpaddingpaddingpaddingpaddingpaddingpaddingpadding
 </style>
